Поправил ошибку с equals, hashCode, toString
