+++ src/animals/Amphibians.java
@@ -1,5 +1,7 @@
 package animals;
 
+import java.util.Objects;
+
 public class Amphibians extends Animals{
 
     public Amphibians(String nickname, int ageYears, String livingEnvironment) {
@@ -21,4 +23,27 @@ public class Amphibians extends Animals{
     protected void move(){
         System.out.println("Земноводные умеют передвигаться и по вде и по суше");
     };
+
+    @Override
+    public String toString() {
+        return "Amphibians{" +
+                "nickname='" + nickname + '\'' +
+                ", ageYears=" + ageYears +
+                ", livingEnvironment='" + livingEnvironment + '\'' +
+                '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        if (!super.equals(o)) return false;
+        return true;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(super.hashCode());
+    }
+
 }
