+++ src/animals/Animals.java
@@ -54,15 +54,6 @@ public abstract class Animals {
     }
 
     @Override
-    public String toString() {
-        return "Animals{" +
-                "nickname='" + nickname + '\'' +
-                ", ageYears=" + ageYears +
-                ", livingEnvironment='" + livingEnvironment + '\'' +
-                '}';
-    }
-
-    @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
