+++ src/animals/Mammals.java
@@ -45,17 +45,6 @@ public abstract class Mammals extends Animals{
     protected abstract void move();
 
     @Override
-    public String toString() {
-        return "Mammals{" +
-                "speed=" + speed +
-                ", foodType='" + foodType + '\'' +
-                ", nickname='" + nickname + '\'' +
-                ", ageYears=" + ageYears +
-                ", livingEnvironment='" + livingEnvironment + '\'' +
-                '}';
-    }
-
-    @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
@@ -68,4 +57,5 @@ public abstract class Mammals extends Animals{
     public int hashCode() {
         return Objects.hash(super.hashCode(), speed, foodType);
     }
+
 }
